+++ stac_fastapi/sqlalchemy/tests/api/test_api.py
@@ -19,7 +19,7 @@ STAC_TRANSACTION_ROUTES = [
     "POST /collections",
     "POST /collections/{collection_id}/items",
     "PUT /collections",
-    "PUT /collections/{collection_id}/items",
+    "PUT /collections/{collection_id}/items/{item_id}",
 ]
 
 
@@ -66,7 +66,9 @@ def test_app_transaction_extension(app_client, load_test_data):
 
 def test_app_search_response(load_test_data, app_client, postgres_transactions):
     item = load_test_data("test_item.json")
-    postgres_transactions.create_item(item, request=MockStarletteRequest)
+    postgres_transactions.create_item(
+        item["collection"], item, request=MockStarletteRequest
+    )
 
     resp = app_client.get("/search", params={"collections": ["test-collection"]})
     assert resp.status_code == 200
@@ -82,7 +84,9 @@ def test_app_search_response_multipolygon(
     load_test_data, app_client, postgres_transactions
 ):
     item = load_test_data("test_item_multipolygon.json")
-    postgres_transactions.create_item(item, request=MockStarletteRequest)
+    postgres_transactions.create_item(
+        item["collection"], item, request=MockStarletteRequest
+    )
 
     resp = app_client.get("/search", params={"collections": ["test-collection"]})
     assert resp.status_code == 200
@@ -96,7 +100,9 @@ def test_app_search_response_geometry_null(
     load_test_data, app_client, postgres_transactions
 ):
     item = load_test_data("test_item_geometry_null.json")
-    postgres_transactions.create_item(item, request=MockStarletteRequest)
+    postgres_transactions.create_item(
+        item["collection"], item, request=MockStarletteRequest
+    )
 
     resp = app_client.get("/search", params={"collections": ["test-collection"]})
     assert resp.status_code == 200
@@ -109,7 +115,9 @@ def test_app_search_response_geometry_null(
 
 def test_app_context_extension(load_test_data, app_client, postgres_transactions):
     item = load_test_data("test_item.json")
-    postgres_transactions.create_item(item, request=MockStarletteRequest)
+    postgres_transactions.create_item(
+        item["collection"], item, request=MockStarletteRequest
+    )
 
     resp = app_client.get("/search", params={"collections": ["test-collection"]})
     assert resp.status_code == 200
@@ -120,7 +128,9 @@ def test_app_context_extension(load_test_data, app_client, postgres_transactions
 
 def test_app_fields_extension(load_test_data, app_client, postgres_transactions):
     item = load_test_data("test_item.json")
-    postgres_transactions.create_item(item, request=MockStarletteRequest)
+    postgres_transactions.create_item(
+        item["collection"], item, request=MockStarletteRequest
+    )
 
     resp = app_client.get("/search", params={"collections": ["test-collection"]})
     assert resp.status_code == 200
@@ -130,7 +140,9 @@ def test_app_fields_extension(load_test_data, app_client, postgres_transactions)
 
 def test_app_query_extension_gt(load_test_data, app_client, postgres_transactions):
     test_item = load_test_data("test_item.json")
-    postgres_transactions.create_item(test_item, request=MockStarletteRequest)
+    postgres_transactions.create_item(
+        test_item["collection"], test_item, request=MockStarletteRequest
+    )
 
     params = {"query": {"proj:epsg": {"gt": test_item["properties"]["proj:epsg"]}}}
     resp = app_client.post("/search", json=params)
@@ -141,7 +153,9 @@ def test_app_query_extension_gt(load_test_data, app_client, postgres_transaction
 
 def test_app_query_extension_gte(load_test_data, app_client, postgres_transactions):
     test_item = load_test_data("test_item.json")
-    postgres_transactions.create_item(test_item, request=MockStarletteRequest)
+    postgres_transactions.create_item(
+        test_item["collection"], test_item, request=MockStarletteRequest
+    )
 
     params = {"query": {"proj:epsg": {"gte": test_item["properties"]["proj:epsg"]}}}
     resp = app_client.post("/search", json=params)
@@ -160,7 +174,9 @@ def test_app_query_extension_limit_lt0(
     load_test_data, app_client, postgres_transactions
 ):
     item = load_test_data("test_item.json")
-    postgres_transactions.create_item(item, request=MockStarletteRequest)
+    postgres_transactions.create_item(
+        item["collection"], item, request=MockStarletteRequest
+    )
 
     params = {"limit": -1}
     resp = app_client.post("/search", json=params)
@@ -171,7 +187,9 @@ def test_app_query_extension_limit_gt10000(
     load_test_data, app_client, postgres_transactions
 ):
     item = load_test_data("test_item.json")
-    postgres_transactions.create_item(item, request=MockStarletteRequest)
+    postgres_transactions.create_item(
+        item["collection"], item, request=MockStarletteRequest
+    )
 
     params = {"limit": 10001}
     resp = app_client.post("/search", json=params)
@@ -182,7 +200,9 @@ def test_app_query_extension_limit_10000(
     load_test_data, app_client, postgres_transactions
 ):
     item = load_test_data("test_item.json")
-    postgres_transactions.create_item(item, request=MockStarletteRequest)
+    postgres_transactions.create_item(
+        item["collection"], item, request=MockStarletteRequest
+    )
 
     params = {"limit": 10000}
     resp = app_client.post("/search", json=params)
@@ -194,7 +214,9 @@ def test_app_sort_extension(load_test_data, app_client, postgres_transactions):
     item_date = datetime.strptime(
         first_item["properties"]["datetime"], "%Y-%m-%dT%H:%M:%SZ"
     )
-    postgres_transactions.create_item(first_item, request=MockStarletteRequest)
+    postgres_transactions.create_item(
+        first_item["collection"], first_item, request=MockStarletteRequest
+    )
 
     second_item = load_test_data("test_item.json")
     second_item["id"] = "another-item"
@@ -202,7 +224,9 @@ def test_app_sort_extension(load_test_data, app_client, postgres_transactions):
     second_item["properties"]["datetime"] = another_item_date.strftime(
         "%Y-%m-%dT%H:%M:%SZ"
     )
-    postgres_transactions.create_item(second_item, request=MockStarletteRequest)
+    postgres_transactions.create_item(
+        second_item["collection"], second_item, request=MockStarletteRequest
+    )
 
     params = {
         "collections": [first_item["collection"]],
@@ -217,7 +241,9 @@ def test_app_sort_extension(load_test_data, app_client, postgres_transactions):
 
 def test_search_invalid_date(load_test_data, app_client, postgres_transactions):
     item = load_test_data("test_item.json")
-    postgres_transactions.create_item(item, request=MockStarletteRequest)
+    postgres_transactions.create_item(
+        item["collection"], item, request=MockStarletteRequest
+    )
 
     params = {
         "datetime": "2020-XX-01/2020-10-30",
@@ -230,7 +256,9 @@ def test_search_invalid_date(load_test_data, app_client, postgres_transactions):
 
 def test_search_point_intersects(load_test_data, app_client, postgres_transactions):
     item = load_test_data("test_item.json")
-    postgres_transactions.create_item(item, request=MockStarletteRequest)
+    postgres_transactions.create_item(
+        item["collection"], item, request=MockStarletteRequest
+    )
 
     point = [150.04, -33.14]
     intersects = {"type": "Point", "coordinates": point}
@@ -247,7 +275,9 @@ def test_search_point_intersects(load_test_data, app_client, postgres_transactio
 
 def test_datetime_non_interval(load_test_data, app_client, postgres_transactions):
     item = load_test_data("test_item.json")
-    postgres_transactions.create_item(item, request=MockStarletteRequest)
+    postgres_transactions.create_item(
+        item["collection"], item, request=MockStarletteRequest
+    )
     alternate_formats = [
         "2020-02-12T12:30:22+00:00",
         "2020-02-12T12:30:22.00Z",
@@ -269,7 +299,9 @@ def test_datetime_non_interval(load_test_data, app_client, postgres_transactions
 
 def test_bbox_3d(load_test_data, app_client, postgres_transactions):
     item = load_test_data("test_item.json")
-    postgres_transactions.create_item(item, request=MockStarletteRequest)
+    postgres_transactions.create_item(
+        item["collection"], item, request=MockStarletteRequest
+    )
 
     australia_bbox = [106.343365, -47.199523, 0.1, 168.218365, -19.437288, 0.1]
     params = {
@@ -286,7 +318,9 @@ def test_search_line_string_intersects(
     load_test_data, app_client, postgres_transactions
 ):
     item = load_test_data("test_item.json")
-    postgres_transactions.create_item(item, request=MockStarletteRequest)
+    postgres_transactions.create_item(
+        item["collection"], item, request=MockStarletteRequest
+    )
 
     line = [[150.04, -33.14], [150.22, -33.89]]
     intersects = {"type": "LineString", "coordinates": line}
@@ -305,7 +339,9 @@ def test_app_fields_extension_return_all_properties(
     load_test_data, app_client, postgres_transactions
 ):
     item = load_test_data("test_item.json")
-    postgres_transactions.create_item(item, request=MockStarletteRequest)
+    postgres_transactions.create_item(
+        item["collection"], item, request=MockStarletteRequest
+    )
 
     resp = app_client.get(
         "/search", params={"collections": ["test-collection"], "fields": "properties"}
@@ -323,7 +359,9 @@ def test_app_fields_extension_return_all_properties(
 
 def test_landing_forwarded_header(load_test_data, app_client, postgres_transactions):
     item = load_test_data("test_item.json")
-    postgres_transactions.create_item(item, request=MockStarletteRequest)
+    postgres_transactions.create_item(
+        item["collection"], item, request=MockStarletteRequest
+    )
 
     response = app_client.get(
         "/",
@@ -341,7 +379,9 @@ def test_app_search_response_forwarded_header(
     load_test_data, app_client, postgres_transactions
 ):
     item = load_test_data("test_item.json")
-    postgres_transactions.create_item(item, request=MockStarletteRequest)
+    postgres_transactions.create_item(
+        item["collection"], item, request=MockStarletteRequest
+    )
 
     resp = app_client.get(
         "/search",
@@ -357,7 +397,9 @@ def test_app_search_response_x_forwarded_headers(
     load_test_data, app_client, postgres_transactions
 ):
     item = load_test_data("test_item.json")
-    postgres_transactions.create_item(item, request=MockStarletteRequest)
+    postgres_transactions.create_item(
+        item["collection"], item, request=MockStarletteRequest
+    )
 
     resp = app_client.get(
         "/search",
@@ -376,7 +418,9 @@ def test_app_search_response_duplicate_forwarded_headers(
     load_test_data, app_client, postgres_transactions
 ):
     item = load_test_data("test_item.json")
-    postgres_transactions.create_item(item, request=MockStarletteRequest)
+    postgres_transactions.create_item(
+        item["collection"], item, request=MockStarletteRequest
+    )
 
     resp = app_client.get(
         "/search",